Add PeopleService to retrieve people data using Apollo
This commit adds a new service called PeopleService that uses Apollo to retrieve data from a GraphQL API. The getPeople() method sends a query to the API to fetch a list of people and returns an Observable of Person objects. The Person interface is also created.

The PeopleService is used in a component called PeopleComponent that retrieves the list of people and stores them in the people property using the OnInit lifecycle hook. This component can be used to display the list of people in the UI.