Implement searchPerson function in PeopleService
This commit adds the implementation of the searchPerson function in the PeopleService. The function sends a GraphQL query to search for a person by name and returns an observable of Person array. The query takes in a variable name and returns the name, height, mass, gender, and homeworld of the person. The query is executed using the apollo.watchQuery function from the apollo-angular library. The result is then mapped to return the data of the searchPerson query.