Extract GraphQL queries to constants file
This commit extracts the GraphQL queries used in the PeopleService to a separate constants file, "queries.ts". This makes the code more maintainable and easier to read. The queries are now imported into the PeopleService file and used as constants. This commit does not change the functionality of the application, only the structure of the code.