Add nginx configurations and Dockerfile for containerization
In this commit, I added nginx configurations that will be used when running the application inside a Docker container. I also included a Dockerfile that will be used to build the Docker image. These configurations are important in ensuring that the application runs smoothly and is accessible through the desired port.